Refactor CommandKeyMap
* follow similar pattern to keyboardLayout
* all user configurations accessible through configuration class
* simplified the custom keyboard layout class
