+++ src/operator/delete.ts
@@ -13,7 +13,7 @@ export class DeleteOperator extends Operator {
         super(modeHandler);
     }
 
-    public key(): string { return "d"; }
+    // public key(): string { return "d"; }
 
     /**
      * Run this operator on a range.
